backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
openai
streamlit
plotly
python-dotenv
requests
google-generativeai